Extract pure utilities from launch.py into launch_helpers.py
Keeps launch.py focused on CLI/orchestration (Args dataclass, render_student/
render_advisor, main loop). Moves the reusable, Args-free helpers into
launch_helpers.py so they can be pulled in by future scripts (teardown,
status) and unit-tested in isolation.

No behavior change — dry-run renders byte-for-byte the same manifests.

Moved: render_template, render_configmap, target_repo_slug, _load_dotenv,
resolve_github_token, render_token_secret,
preflight_check_target_repo_access, kubectl_apply, STUDENT_NAMES.

resolve_github_token now takes the dotenv path as an explicit argument
instead of computing it relative to its own module location, so the
helper doesn't embed an assumption about its callers' layout.
